requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask-migrate
alembic
sqlalchemy_schemadisplay 
graphviz
faker
Flask-Cors
requests